requirements.txt
appdirs==1.4.3
argcomplete==1.9.4
Django==2.0.3
Pillow==5.1.0
psycopg2==2.7.4
pyserial==3.4
python-escpos==2.2.0
pytz==2018.3
pyusb==1.0.2
PyYAML==3.12
qrcode==6.0
six==1.11.0